光谱顺序问题修正

+++ USB_Host_Library/Class/USB4000/Inc/usbh_template.h
@@ -92,8 +92,8 @@ typedef struct _USB4000_Process
     int16_t*             pin_Spectrum;              //光谱数组
     int32_t*             pl_SumSpectrum;            //求和光谱
     
-    double*              plf_ProcessSpectrum;       //处理后的光谱
-    double*              plf_WaveLenth;             //波长数组
+    float*              plf_ProcessSpectrum;       //处理后的光谱
+    float*              plf_WaveLenth;             //波长数组
     uint16_t             uin_Pixels;                //像素个数
     
     uint8_t              uch_ScansToAverage;        //多次扫描平均
